Reorganize component structure
