+++ src/app/admin/dashboard/page.tsx
@@ -2,14 +2,12 @@
 const baseUrl = process.env.NEXT_PUBLIC_LOCAL_URL || "";
 
 import { useState, useEffect, useCallback } from "react";
-import Image from "next/image"; 
 import { useRouter } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import { Bell, Menu } from "lucide-react";
 import toast from "react-hot-toast";
 
 // Modular Imports
-import { AdminSidebar } from "@/components/admin/AdminSidebar";
+import { AdminSidebar } from "@/components/layout/AdminSidebar";
+import { AdminDashboardHeader } from "@/components/layout/AdminDashboardHeader";
 import { VerificationTab } from "@/components/admin/tabs/VerificationTab"; // Original Admin Tab
 import { ProfileTab } from "@/components/admin/tabs/ProfileTab";
 import { MasterlistTab } from "@/components/admin/tabs/MasterlistTab";
@@ -197,45 +195,7 @@ export default function AdminDashboard() {
       {/* FIX APPLIED HERE: md:p-8 was changed to md:px-8 md:pt-4 */}
       <main className="flex-1 md:ml-72 p-4 md:px-8 md:pt-4 min-h-screen bg-[#FDFBF7]">
         
-        <header className="flex items-center justify-between mb-8 py-4 border-b border-stone-200/50">
-            <div className="flex items-center gap-4">
-                <Button variant="ghost" size="icon" className="md:hidden text-stone-500" onClick={() => setIsMobileMenuOpen(true)}>
-                    <Menu className="h-6 w-6"/>
-                </Button>
-
-                <div className="flex items-center gap-3 md:hidden">
-                    <div className="relative w-8 h-8 overflow-hidden hover:scale-105 transition-transform duration-300">
-                        <Image src="/images/umtc-logo.png" alt="UMTC" fill className="object-contain" />
-                    </div>
-                    <div className="h-8 w-[1px] bg-stone-300"></div>
-                    <div className="relative w-8 h-8 overflow-hidden hover:scale-105 transition-transform duration-300">
-                        <Image src="/images/aurium-logo.png" alt="Aurium" fill className="object-contain" />
-                    </div>
-                </div>
-
-                {/* Dynamic Title */}
-                <h1 className="text-2xl font-serif font-bold text-stone-800 hidden md:block">
-                    {activeTab === 'verification' && "Verification Queue"}
-                    {activeTab === 'graduate-review' && "Graduate Verification"}
-                    {activeTab === 'notes' && "Staff Notes"}
-                    {activeTab === 'slots' && "Schedule Manager"}
-                    {activeTab === 'masterlist' && "Verified Masterlist"}
-                    {activeTab === 'scanner' && "Attendance Scanner"}
-                    {activeTab === 'profile' && "My Profile"}
-                    {activeTab === 'roles' && "Manage Staffs"}
-                </h1>
-            </div>
-
-            <div className="flex items-center gap-3">
-               <div className="hidden md:flex items-center gap-2 px-3 py-1 bg-white border border-stone-200 rounded-full shadow-sm text-xs font-medium text-stone-500">
-                  <div className="w-2 h-2 rounded-full bg-green-500 animate-pulse"></div> System Online
-               </div>
-               <Button variant="ghost" size="icon" className="relative rounded-full text-stone-400 hover:text-amber-800">
-                   <Bell className="h-5 w-5"/>
-                   <span className="absolute top-2 right-2 h-2 w-2 bg-red-500 rounded-full border-2 border-white"></span>
-               </Button>
-            </div>
-        </header>
+        <AdminDashboardHeader activeTab={activeTab} onOpenMenu={() => setIsMobileMenuOpen(true)} />
 
         {/* CONTENT AREA */}
         <div className={`space-y-6 animate-in fade-in slide-in-from-bottom-4 duration-500 ${activeTab !== 'profile' ? 'max-w-7xl mx-auto' : ''}`}>
